[kernel] Build everything with -Wconversion [part 8]
Changes to be committed:
	modified:   include/tilck/common/color_defs.h
	modified:   include/tilck/kernel/timer.h
	modified:   kernel/arch/generic_x86/pit.c
	modified:   kernel/arch/generic_x86/rtc.c
	modified:   kernel/arch/generic_x86/textmode_video.c
	modified:   kernel/arch/generic_x86/uart.c

+++ include/tilck/common/color_defs.h
@@ -39,7 +39,7 @@ enum vga_color {
  * Entry defs (color + char): the hardware format (VGA textmode) is used also
  * by the hw-independents "term" and "fb_console" for convenience.
  */
-#define make_vgaentry(c, color) (((u16)c) | ((u16)color << 8))
+#define make_vgaentry(c, color) ((u16)(((u16)c) | ((u16)color << 8)))
 #define vgaentry_get_fg(e) (((e) >> 8) & 0xF)
 #define vgaentry_get_bg(e) (((e) >> 12) & 0xF)
 #define vgaentry_get_char(e) ((e) & 0xFF)
